+++ yii2-adapter/legacy/base/ConfigurableComponent.php
@@ -16,7 +16,7 @@ use CraftCms\Cms\Support\Facades\Deprecator;
 use CraftCms\Cms\Support\Facades\HtmlStack;
 use CraftCms\Cms\Support\Facades\InputNamespace;
 use CraftCms\Cms\Support\Html;
-use CraftCms\Yii2Adapter\Cp\FormDefinitions\Elements\LegacySettings;
+use CraftCms\Yii2Adapter\Cp\Components\LegacySettings;
 use DateTime;
 use ReflectionClass;
 use ReflectionProperty;